Feature: init target_tools
